split between sampler.py and samples.py
Necessary to avoid circular imports
